+++ lib/parse/api/config.rb
@@ -7,7 +7,7 @@ module Parse
     #object fetch methods
     module Config
       attr_accessor :config
-
+      CONFIG_PATH = "config"
       def config!
         @config = nil
         self.config
@@ -15,13 +15,23 @@ module Parse
 
       def config
         if @config.nil?
-          response = request :get, "config"
+          response = request :get, CONFIG_PATH
           unless response.error?
             @config = response.result["params"]
           end
         end
         @config
       end
+
+      def update_config(params)
+        body = { params: params }
+        response = request :put, CONFIG_PATH, body: body
+        return false if response.error?
+        result = response.result["result"]
+        @config.merge!(params) if result && @config.present?
+        result
+      end
+
     end
 
   end